requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4
supabase==2.3.4

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# AI/ML
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
google-generativeai>=0.7.0

# File handling
python-magic==0.4.27

# Utilities
python-dotenv==1.0.1
httpx>=0.24.0,<0.26.0
tenacity==8.2.3
structlog==24.1.0
aiofiles==23.2.1
numpy==1.26.3

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
